+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v1.2.4
 
 - **remove**: `navigation.courseGreatCircle.nextPoint.bearingTrue` calculation (available in `signalk-derived-data` plugin `Course Data`).
+- **update**: Use `setPluginStatus` and `setPluginError` instead of deprecated `setPoviderStatus` and `setProviderError`
 
 ### v1.2.3